Cast top and bottom slices to uint8
Force regionprops to interpret the slices as a label
matrix defining a single region (all pixels are 0 or 1).

If regionprops gets a logical array instead, it will interpret
discontinuous areas as separate regions.

+++ MATLAB/calcLength.m
@@ -7,8 +7,8 @@ function [len, bottom, top] = calcLength(bw)
   bottom_idx = min(height);
   top_idx = max(height);
 
-  bottom_slice = bw(:, :, bottom_idx);
-  top_slice = bw(:, :, top_idx);
+  bottom_slice = uint8(bw(:, :, bottom_idx));
+  top_slice = uint8(bw(:, :, top_idx));
 
   p_bottom = regionprops(bottom_slice, 'Centroid');
   p_top = regionprops(top_slice, 'Centroid');
